requirements.txt
fastapi
uvicorn
python-dotenv
httpx
starlette
nicegui
